pyproject.toml
[tool.poetry]
name = "search-vacancy-with-sql"
version = "0.1.0"
description = ""
authors = ["maksim <m.merkulovdodo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2 = "^2.9.10"
pytest-postgresql = "^6.1.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"


[tool.poetry.group.pytest.dependencies]
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '(venv|tests)'